bus - separated local and distributed bus implementations
